requirements.txt
discord.py>=2.3.0
fastapi>=0.104.0
python-dotenv>=0.21.0
uvicorn>=0.24.0
psycopg2-binary>=2.9.9
pydantic>=2.4.0
pydantic-settings>=2.0.0
mypy>=1.7.0
sqlalchemy>=2.0.36

